Cleaning up util usage and test setup
