+++ test/contracts/ExchangeState.js
@@ -2,14 +2,11 @@
 
 const { artifacts, contract } = require('@nomiclabs/buidler');
 
-const { assert } = require('../utils/common');
+const { assert } = require('./common');
 
 const { toBytes32 } = require('../..');
-const { toUnit } = require('../utils/testUtils');
-const {
-	onlyGivenAddressCanInvoke,
-	ensureOnlyExpectedMutativeFunctions,
-} = require('../utils/setupUtils');
+const { toUnit } = require('../utils');
+const { onlyGivenAddressCanInvoke, ensureOnlyExpectedMutativeFunctions } = require('./helpers');
 
 const { isBN } = require('web3-utils');
