+++ test/contracts/FeePoolState.js
@@ -2,13 +2,10 @@
 
 const { artifacts, contract } = require('@nomiclabs/buidler');
 
-const { assert } = require('../utils/common');
+const { assert } = require('./common');
 
-const { toPreciseUnit, toUnit } = require('../utils/testUtils');
-const {
-	onlyGivenAddressCanInvoke,
-	ensureOnlyExpectedMutativeFunctions,
-} = require('../utils/setupUtils');
+const { toPreciseUnit, toUnit } = require('../utils');
+const { onlyGivenAddressCanInvoke, ensureOnlyExpectedMutativeFunctions } = require('./helpers');
 
 const FeePoolState = artifacts.require('FeePoolState');
