+++ test/contracts/helpers.js
@@ -4,8 +4,8 @@ const abiDecoder = require('abi-decoder');
 
 const { assert } = require('./common');
 
-const { currentTime, toUnit } = require('../utils/testUtils');
-const { toBytes32 } = require('../../.');
+const { currentTime, toUnit } = require('../utils');
+const { toBytes32 } = require('../..');
 
 module.exports = {
 	/**
